Update the base Task model Move generic task views and actions into the base task model. Single answer and multiple choice tasks now only need to define their respective default annotations.
