+++ packages/lib-classifier/src/store/tasks/MultipleChoiceTask.js
@@ -16,22 +16,8 @@ const MultipleChoice = types.model('MultipleChoice', {
   type: types.literal('multiple')
 })
 .views(self => ({
-  get annotation () {
-    const { currentAnnotations } = getRoot(self).classifications
-    let currentAnnotation
-    if (currentAnnotations && currentAnnotations.size > 0) {
-      currentAnnotation = currentAnnotations.get(self.taskKey)
-    }
-    return currentAnnotation || self.defaultAnnotation
-  },
   get defaultAnnotation () {
     return MultipleChoiceAnnotation.create({ task: self.taskKey })
-  }
-}))
-.actions(self => ({
-  updateAnnotation (value) {
-    const { addAnnotation } = getRoot(self).classifications
-    addAnnotation(value, self)
   }
 }))
