+++ packages/lib-classifier/src/store/tasks/SingleChoiceTask.js
@@ -1,4 +1,4 @@
-import { getRoot, types } from 'mobx-state-tree'
+import { types } from 'mobx-state-tree'
 import Task from './Task'
 import SingleChoiceAnnotation from '../annotations/SingleChoiceAnnotation'
 
@@ -16,22 +16,8 @@ const SingleChoice = types.model('SingleChoice', {
   type: types.literal('single')
 })
 .views(self => ({
-  get annotation () {
-    const { currentAnnotations } = getRoot(self).classifications
-    let currentAnnotation
-    if (currentAnnotations && currentAnnotations.size > 0) {
-      currentAnnotation = currentAnnotations.get(self.taskKey)
-    }
-    return currentAnnotation || self.defaultAnnotation
-  },
   get defaultAnnotation () {
     return SingleChoiceAnnotation.create({ task: self.taskKey })
-  }
-}))
-.actions(self => ({
-  updateAnnotation (value) {
-    const { addAnnotation } = getRoot(self).classifications
-    addAnnotation(value, self)
   }
 }))
