+++ packages/lib-classifier/src/store/tasks/Task.js
@@ -1,7 +1,28 @@
-import { types } from 'mobx-state-tree'
+import { getRoot, types } from 'mobx-state-tree'
+import Annotation from '../annotations/Annotation'
 
 const Task = types.model('Task', {
   taskKey: types.identifier
 })
+.views(self => ({
+  get annotation () {
+    const { currentAnnotations } = getRoot(self).classifications
+    let currentAnnotation
+    if (currentAnnotations && currentAnnotations.size > 0) {
+      currentAnnotation = currentAnnotations.get(self.taskKey)
+    }
+    return currentAnnotation || self.defaultAnnotation
+  },
+  get defaultAnnotation () {
+    // Override this in a real task
+    return Annotation.create({ task: self.taskKey })
+  }
+}))
+.actions(self => ({
+  updateAnnotation (value) {
+    const { addAnnotation } = getRoot(self).classifications
+    addAnnotation(value, self)
+  }
+}))
 
 export default Task
